LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-volume-controls-hidden-when-mute-button-disabled.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the volume controls aren't visible when the mute button is disabled.");

const mediaControls = new MacOSFullscreenMediaControls({ width: 250, height: 250 });
const leftContainer = mediaControls.bottomControlsBar.children[1];

debug("Mute button is enabled");
shouldBeTrue("leftContainer.visible");

debug("")
debug("Mute button is disabled");
mediaControls.muteButton.enabled = false;
shouldBeFalse("leftContainer.visible");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
