LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-volume-controls-hidden-when-mute-button-disabled-expected.txt

Testing the volume controls aren't visible when the mute button is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Mute button is enabled
PASS leftContainer.visible is true

Mute button is disabled
PASS leftContainer.visible is false

PASS successfullyParsed is true

TEST COMPLETE
